requirements.txt
protobuf==4.25.3
langchain==0.3.25
langchain_chroma==0.2.4
langchain_core==0.3.65
langchain_tavily==0.2.2
langchain_text_splitters==0.3.8
pandas==2.3.0
pdfplumber==0.11.6
Pillow==11.2.1
python_docx==1.1.2
Requests==2.32.4
streamlit==1.45.1
streamlit_lottie==0.0.5
google-generativeai
google-ai-generativelanguage==0.6.15
langchain_google_genai
python-dotenv
